requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
Flask==2.2.2
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
